Generalise CircularBuffer.
